requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.10
# by the following command:
#
#    pip-compile --no-emit-index-url requirements.in
#

allure-pytest==2.13.1
    # via -r requirements.in
allure-python-commons==2.13.1
    # via
    #   -r requirements.in
    #   allure-pytest
attrs==22.2.0
    # via
    #   allure-python-commons
    #   pytest
exceptiongroup==1.1.1
    # via pytest
iniconfig==2.0.0
    # via pytest
packaging==23.0
    # via pytest
pluggy==1.0.0
    # via
    #   allure-python-commons
    #   pytest
py==1.11.0
    # via pytest-html
pytest==7.2.2
    # via
    #   -r requirements.in
    #   allure-pytest
    #   pytest-html
    #   pytest-metadata
pytest-html==3.2.0
    # via -r requirements.in
pytest-metadata==2.0.4
    # via pytest-html
tomli==2.0.1
    # via pytest
